Widen island serials from 16 to 22 bits
Body entities packed the island serial into 16 bits. Serials are monotonic and
never reused, so the space is consumed by *cumulative* body creation rather
than by how many bodies are live -- and past 65,536 every new body aliases onto
a live one. A reported session was at 4,389 bodies / 4,224 islands / 17,143
broken bonds, which is the regime where that ceiling is reachable.

Rebalanced the 28 usable bits: 6 for the structure (64, of which we place 16
and never will place 64) and 22 for the serial (4.19M, up from 65k). Changed in
lockstep on both sides of the wire -- destruction/src/ids.rs and
physx-bridge/src/destruction.cc pack it, client/src/city/topology.ts unpacks
it -- along with every u16 serial type behind them.

This does NOT fix the duplicate-body-id bug, and the commit should not be read
as claiming otherwise. At the reproducing scale duplicates went 4423 -> 651
across separate runs, but the bench never wraps the serial counter at all (the
allocator is instrumented to report exhaustion and stays silent), so that
movement is run-to-run variance, not a fix. The aliasing at bench scale has a
different cause, still open, and the regression assertion still fails.

What this does buy: it removes a ceiling that a long session with heavy
destruction can genuinely hit, and it takes serial exhaustion off the table as
a confound while the real cause is found.

Verified: 106 Rust tests and 23 client tests pass; body_entity(0, 1) is still
0x8000_0001 so the pinned wire expectation is unchanged.

+++ destruction/src/encoder.rs
@@ -236,7 +236,7 @@ impl ChunkStreamEncoder {
         sim_tick: u32,
         active: &[BodySnapshotInput],
         output: &DestructionTickOutput,
-        wakes: &[(u32, u16)],
+        wakes: &[(u32, u32)],
     ) {
         // Ledger + staged topology.
         if !output.batches.is_empty() || !output.settled.is_empty() || !wakes.is_empty() {
@@ -244,7 +244,7 @@ impl ChunkStreamEncoder {
                 self.ledger.apply_batch(batch);
                 for promotion in &batch.promoted_islands {
                     let entity =
-                        ids::body_entity(batch.structure_id, promotion.island_id as u16);
+                        ids::body_entity(batch.structure_id, u32::from(promotion.island_id));
                     let nodes: Vec<u32> = promotion
                         .chunks
                         .iter()
@@ -265,7 +265,7 @@ impl ChunkStreamEncoder {
                     );
                 }
                 for &retired in &batch.retired_island_ids {
-                    let entity = ids::body_entity(batch.structure_id, retired as u16);
+                    let entity = ids::body_entity(batch.structure_id, retired);
                     self.bodies.remove(&entity);
                     for client in self.clients.values_mut() {
                         client.tracks.remove(&entity);
@@ -277,7 +277,7 @@ impl ChunkStreamEncoder {
             }
             for settle in &output.settled {
                 self.ledger.apply_settle(settle);
-                let entity = ids::body_entity(settle.structure_id, settle.island_id as u16);
+                let entity = ids::body_entity(settle.structure_id, u32::from(settle.island_id));
                 if let Some(track) = self.bodies.get_mut(&entity) {
                     track.settled_hint = true;
                 }
